[Refactoring] Introduce Remove component
Responsible for displaying a confirmation box on remove
