+++ app/components/AnnotationForm/presenter.jsx
@@ -1,6 +1,6 @@
 import React, { PropTypes, Component } from 'react';
 import Immutable from 'immutable';
-import Remove from './Remove';
+import Remove from '../Remove';
 import InputNumber from '../InputNumber';
 
 const { array, number, string, shape, func, instanceOf } = PropTypes;
@@ -190,8 +190,10 @@ class AnnotationForm extends Component {
         {this.state.displayRemoveForm ?
           <Remove
             onRemove={this.handleRemove}
-            onActionRemoveCancelClick={this.toggleActionRemove}
-          /> : null
+            onCancel={this.toggleActionRemove}
+          >
+            Are you sure you want to remove this annotation?
+          </Remove> : null
         }
       </div>
     );
