+++ app/scss/components/_remove.scss
@@ -1,0 +1,29 @@
+/*
+ * Remove component
+ */
+
+.action-remove {
+  border-radius: $global-radius;
+  background-color: lighten($light-yeah, 10%);
+  color: $white;
+
+  button {
+    margin-bottom: 0 !important;
+
+    &.remove {
+      @include button(
+        $background: $alert-color,
+        $background-hover: darken($alert-color, 10%)
+      );
+      margin-right: 0.5rem;
+    }
+
+    &.cancel {
+      @include button(
+        $background: $grey,
+        $background-hover: darken($grey, 10%),
+        $color: $body-font-color
+      );
+    }
+  }
+}
