+++ app/scss/main.scss
@@ -36,3 +36,4 @@ $fa-font-path: '~font-awesome/fonts';
 @import 'components/annotations';
 @import 'components/exons';
 @import 'components/search';
+@import 'components/remove';
